+++ common/model/src/main/java/is/codion/common/model/condition/TableConditionModel.java
@@ -67,7 +67,7 @@ public interface TableConditionModel<C> {
 	/**
 	 * @return an observer notified each time the condition changes
 	 */
-	Observer<?> conditionChanged();
+	Observer<?> changed();
 
 	/**
 	 * Instantiates a new {@link TableConditionModel}