+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingForeignKeyConditionModel.java
@@ -156,8 +156,8 @@ public final class SwingForeignKeyConditionModel implements ConditionModel<Attri
 	}
 
 	@Override
-	public Observer<?> conditionChanged() {
-		return condition.conditionChanged();
+	public Observer<?> changed() {
+		return condition.changed();
 	}
 
 	/**